+++ packages/document-grounding/src/tests/retrieval-api.test.ts
@@ -3,12 +3,10 @@ import {
   aiCoreDestination,
   mockClientCredentialsGrantCall
 } from '../../../../test-util/mock-http.js';
-import {
-  CollectionsApi,
-  type CollectionsListResponse
-} from '../client/api/index.js';
+import { RetrievalApi } from '../client/api/index.js';
+import type { DataRepositories } from '../../internal.js';
 
-describe('collections', () => {
+describe('retrieval data repository', () => {
   beforeEach(() => {
     mockClientCredentialsGrantCall();
   });
@@ -16,16 +14,14 @@ describe('collections', () => {
     nock.cleanAll();
   });
 
-  it('should get all collections', async () => {
-    const expectedResponse: CollectionsListResponse = {
+  it('should get all data repositories', async () => {
+    const expectedResponse: DataRepositories = {
       resources: [
         {
+          id: '717dc2b3-835b-4725-8854-fa5c7698a416',
           title: 'ai-sdk-js-demo',
-          embeddingConfig: {
-            modelName: 'text-embedding-ada-002-v2'
-          },
           metadata: [],
-          id: '7171c2b2-835b-4745-8824-ff5c7e98a416'
+          type: 'vector'
         }
       ],
       count: 1
@@ -36,16 +32,15 @@ describe('collections', () => {
         'AI-Resource-Group': 'default'
       }
     })
-      .get('/v2/lm/document-grounding/vector/collections')
+      .get('/v2/lm/document-grounding/retrieval/dataRepositories')
       .reply(200, expectedResponse, {
         'Content-Type': 'application/json'
       });
 
-    const result: CollectionsListResponse =
-      await CollectionsApi.vectorV1VectorEndpointsGetAllCollections(
-        {},
-        { 'AI-Resource-Group': 'default' }
-      ).execute();
+    const result: DataRepositories = await RetrievalApi.getDataRepositories(
+      {},
+      { 'AI-Resource-Group': 'default' }
+    ).execute();
 
     expect(result).toEqual(expectedResponse);
   });
